Implement CopyableObjectProxy
This implements the CopyableObjectProxy class as a subclass of
ObjectProxy that implements the logic necessary for using copy.copy()
and copy.deepcopy() on proxied objects. This is currently implemented
as a subclass class because the __copy__() and __deepcopy__() methods
don't work when put directly into the ObjectProxy class for some
reason.